\name{arthritisG}
\docType{data}
\alias{arthritisG}
\title{A genotype data set with two subpopulations and moderate ancestral differences}
\description{
This is a genotype data set containing the genotypes of 868 cases and 1,194 controls at 12,749 SNPs 
that are not associated with the disease. See also \link{arthritisP}.
}
\source{Genetic Analysis Workshop 16 Problem 1}
\references{
Amos CI, Chen WV, Seldin MF, Remmers EF, Taylor KE, Criswell LA, Lee AT, Plenge RM,
Kastner DL, and Gregersen PK. Data for Genetic Analysis Workshop 16 Problem 1,
association analysis of rheumatoid arthritis data. \emph{In BMC proceedings}. 2009; 3: S2.

Yu K, Wang Z, Li Q, Wacholder S, Hunter DJ, Hoover RN, Chanock S, and Thomas G.
Population substructure and control selection in genome-wide association studies. \emph{PloS
one}. 2008; 3(7): e2551.
} 
\usage{data(arthritisG)}
\keyword{datasets}