\name{arthritisP}
\docType{data}
\alias{arthritisP}
\title{A phenotype data set with two subpopulations and moderate ancestral differences}
\description{
This is a phenotype data set consisting of 868 cases and 1,194 controls. 
See also \link{arthritisG}.
}
\source{Genetic Analysis Workshop 16 Problem 1}
\references{
Amos CI, Chen WV, Seldin MF, Remmers EF, Taylor KE, Criswell LA, Lee AT, Plenge RM,
Kastner DL, and Gregersen PK. Data for Genetic Analysis Workshop 16 Problem 1,
association analysis of rheumatoid arthritis data. \emph{In BMC proceedings}. 2009; 3: S2.
}
\usage{data(arthritisP)}
\keyword{datasets}