\name{drG}
\docType{data}
\alias{drG}
\title{A simulation genotype data set for distance regression}
\description{
This is a simulation genotype data set generated using the real data set that contains 
the genotypes of 127 SNPs in the uronyl-2-sulfotransferase gene from Genetic Analysis Workshop 16. After 
deleting the subjects containing the missing value, the data set 
consists of 1,081 subjects. See also \link{drP} and \link{drS}.
}
\usage{data(drG)}
\keyword{datasets}
\references{
Y Lin, M Zhang, L Wang, et al. Simultaneous genome-wide association studies 
of anti-cyclic citrullinated peptide in rheumatoid arthritis using penalized 
orthogonal-components regression. \emph{BioMed Central Ltd}. 2009; 3(Suppl 7): S20.
}