\name{drP}
\docType{data}
\alias{drP}
\title{A simulation phenotype data set for distance regression}
\description{
This is a simulation phenotype data set consisting of 1,081 subjects. 
It corresponds to the data set \link{drG} when the disease 
prevalence is set to be 0.05 and the first 50 SNPs in \link{drG} are assumed to be 
associated with the disease with the log odds ratio \eqn{ln(1.05)}. See also \link{drG} and \link{drS}.
}
\usage{data(drP)}
\keyword{datasets}