\name{drS}
\docType{data}
\alias{drS}
\title{A similarity matrix for distance regression}
\description{
This data set contains the similarity matrix among the subjects 
calculated from the simulation genotype data set \link{drG}, 
which contains the genotypes of 1,081 subjects at 127 markers.
See also \link{drG} and \link{drP}. 
}
\usage{data(drS)}
\keyword{datasets}