\name{extreme2PSG}
\docType{data}
\alias{extreme2PSG}
\title{A simulation genotype data set with two subpopulations and extreme ancestral differences}
\description{
This is a simulation genotype data set containing the genotypes of 1,000 cases and 1,000 controls at 10,000 SNPs 
that are not associated with the disease. The population substructure is S1 
(two underlying discrete subpopulations) and the level of the ancestral differences between cases 
and controls is extreme. The sampling fractions of the two subpopulations are (0.5, 0.5) for cases and 
(0, 1) for controls. See also \link{extreme2PSP}.
}
\usage{data(extreme2PSG)}
\keyword{datasets}