\name{extreme2PSP}
\docType{data}
\alias{extreme2PSP}
\title{A simulation phenotype data set with two subpopulations and extreme ancestral differences}
\description{
This is a simulation phenotype data set consisting of 1,000 cases and 1,000 controls. 
The population substructure is S1 (two underlying discrete subpopulations) 
and the level of the ancestral differences between cases and controls is extreme. 
The sampling fractions of the two subpopulations are (0.5, 0.5) for cases and (0, 1) for controls. 
See also \link{extreme2PSG}.
}
\usage{data(extreme2PSP)}
\keyword{datasets}