\name{moderate2PSP}
\docType{data}
\alias{moderate2PSP}
\title{A simulation phenotype data set with two subpopulations and moderate ancestral differences}
\description{
This is a simulation phenotype data set consisting of 1,000 cases and 1,000 controls. 
The population substructure is S1 
(two underlying discrete subpopulations) and the level of the 
ancestral differences between cases and controls is 
moderate. The sampling fractions of the two subpopulations are (0.6, 0.4) for cases and 
(0.4, 0.6) for controls. See also \link{moderate2PSG}.
}
\usage{data(moderate2PSP)}
\keyword{datasets}