\name{moderate3PSG}
\docType{data}
\alias{moderate3PSG}
\title{A simulation genotype data set with three subpopulations and moderate ancestral differences}
\description{
This is a simulation genotype data set containing the genotypes of 1,000 cases and 1,000 controls at 10,000 SNPs 
that are not associated with the disease. The population substructure is S2 
(three underlying discrete subpopulations) and the level of the ancestral differences between cases 
and controls is moderate. The sampling fractions of the three subpopulations are (0.45, 0.35, 0.20) for cases and 
(0.35, 0.20, 0.45) for controls. See also \link{moderate3PSP}.
}
\usage{data(moderate3PSG)}
\keyword{datasets}