\name{moderate3PSP}
\docType{data}
\alias{moderate3PSP}
\title{A simulation phenotype data set with three subpopulations and moderate ancestral differences}
\description{
This is a simulation phenotype data set consisting of 1,000 cases and 1,000 controls. 
The population substructure is S2 (three underlying discrete subpopulations) 
and the level of the ancestral differences between cases and controls is moderate. 
The sampling fractions of the three subpopulations are (0.45, 0.35, 0.20) for cases and (0.35, 0.20, 0.45) for controls. 
See also \link{moderate3PSG}.
}
\usage{data(moderate3PSP)}
\keyword{datasets}